fix reproduce schema, and make outputs available on reproduce
