+++ frontend/src/api/experiments/Experiments.ts
@@ -3,7 +3,6 @@ import axios from 'utils/axios'
 import { BASE_URL } from 'const/API'
 import { OutputPathsDTO } from 'api/run/Run'
 import { EXPERIMENTS_STATUS } from 'store/slice/Experiments/ExperimentsType'
-import { WorkflowConfigDTO } from 'api/workflow/Workflow'
 
 export type ExperimentsDTO = {
   [uid: string]: ExperimentDTO
@@ -39,8 +38,6 @@ export type ExperimentDTO = {
   hasNWB: boolean
   nwb: NWBType
 }
-
-export type FetchExperimentDTO = ExperimentDTO & WorkflowConfigDTO
 
 export async function getExperimentsApi(
   workspaceId: number,
@@ -96,15 +93,6 @@ export async function downloadExperimentConfigApi(
     {
       responseType: 'blob',
     },
-  )
-  return response.data
-}
-
-export async function fetchExperimentApi(
-  workspace_id: number,
-): Promise<FetchExperimentDTO> {
-  const response = await axios.get(
-    `${BASE_URL}/experiments/fetch/${workspace_id}`,
   )
   return response.data
 }
