+++ frontend/src/store/slice/Pipeline/PipelineSlice.ts
@@ -1,8 +1,8 @@
 import { createSlice, isAnyOf, PayloadAction } from '@reduxjs/toolkit'
-import { fetchExperiment } from '../Experiments/ExperimentsActions'
 import {
   reproduceWorkflow,
   importWorkflowConfig,
+  fetchWorkflow,
 } from 'store/slice/Workflow/WorkflowActions'
 import {
   cancelResult,
@@ -82,40 +82,43 @@ export const pipelineSlice = createSlice({
           status: RUN_STATUS.START_UNINITIALIZED,
         }
       })
-      .addCase(fetchExperiment.rejected, () => initialState)
-      .addCase(fetchExperiment.fulfilled, (state, action) => {
-        state.currentPipeline = {
-          uid: action.payload.unique_id,
-        }
-        state.runBtn = RUN_BTN_OPTIONS.RUN_ALREADY
-        state.run = {
-          uid: action.payload.unique_id,
-          status: RUN_STATUS.START_SUCCESS,
-          runResult: {
-            ...convertToRunResult(
-              convertFunctionsToRunResultDTO(action.payload.function),
-            ),
-          },
-          runPostData: {
-            name: action.payload.name,
-            nodeDict: action.payload.nodeDict,
-            edgeDict: action.payload.edgeDict,
-            snakemakeParam: {},
-            nwbParam: {},
-            forceRunList: [],
-          },
-        }
-
-        const runResultPendingList = Object.values(state.run.runResult).filter(
-          isNodeResultPending,
-        )
-        if (runResultPendingList.length === 0) {
-          state.run.status = RUN_STATUS.FINISHED
-        }
-      })
+      .addCase(fetchWorkflow.rejected, () => initialState)
       .addCase(cancelResult.fulfilled, (state, action) => {
         state.run.status = RUN_STATUS.CANCELED
       })
+      .addMatcher(
+        isAnyOf(fetchWorkflow.fulfilled, reproduceWorkflow.fulfilled),
+        (state, action) => {
+          state.currentPipeline = {
+            uid: action.payload.unique_id,
+          }
+          state.runBtn = RUN_BTN_OPTIONS.RUN_ALREADY
+          state.run = {
+            uid: action.payload.unique_id,
+            status: RUN_STATUS.START_SUCCESS,
+            runResult: {
+              ...convertToRunResult(
+                convertFunctionsToRunResultDTO(action.payload.function),
+              ),
+            },
+            runPostData: {
+              name: action.payload.name,
+              nodeDict: action.payload.nodeDict,
+              edgeDict: action.payload.edgeDict,
+              snakemakeParam: {},
+              nwbParam: {},
+              forceRunList: [],
+            },
+          }
+
+          const runResultPendingList = Object.values(
+            state.run.runResult,
+          ).filter(isNodeResultPending)
+          if (runResultPendingList.length === 0) {
+            state.run.status = RUN_STATUS.FINISHED
+          }
+        },
+      )
       .addMatcher(
         isAnyOf(run.pending, runByCurrentUid.pending),
         (state, action) => {
