+++ frontend/src/store/slice/Workflow/WorkflowActions.ts
@@ -1,13 +1,27 @@
 import { WORKFLOW_SLICE_NAME } from './WorkflowType'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import {
+  fetchWorkflowApi,
   reproduceWorkflowApi,
   importWorkflowConfigApi,
   WorkflowConfigDTO,
+  WorkflowWithResultDTO,
 } from 'api/workflow/Workflow'
 
+export const fetchWorkflow = createAsyncThunk<WorkflowWithResultDTO, number>(
+  `${WORKFLOW_SLICE_NAME}/fetchExperiment`,
+  async (workspaceId, thunkAPI) => {
+    try {
+      const response = await fetchWorkflowApi(workspaceId)
+      return response
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e)
+    }
+  },
+)
+
 export const reproduceWorkflow = createAsyncThunk<
-  WorkflowConfigDTO,
+  WorkflowWithResultDTO,
   { workspaceId: number; uid: string }
 >(
   `${WORKFLOW_SLICE_NAME}/reproduceWorkflow`,
