+++ studio/app/common/core/workflow/workflow.py
@@ -102,9 +102,3 @@ class RunItem(BaseModel):
     snakemakeParam: dict = {}
     nwbParam: dict = {}
     forceRunList: List[ForceRun]
-
-
-@dataclass
-class WorkflowConfig:
-    nodeDict: Dict[str, Node]
-    edgeDict: Dict[str, Edge]
