+++ studio/app/common/routers/experiment.py
@@ -1,6 +1,5 @@
 import os
 import shutil
-from dataclasses import asdict
 from glob import glob
 from typing import Dict
 
@@ -9,18 +8,12 @@ from fastapi.responses import FileResponse
 
 from studio.app.common.core.experiment.experiment import ExptConfig
 from studio.app.common.core.experiment.experiment_reader import ExptConfigReader
-from studio.app.common.core.experiment.experiment_utils import ExptUtils
 from studio.app.common.core.utils.filepath_creater import join_filepath
-from studio.app.common.core.workflow.workflow_reader import WorkflowConfigReader
 from studio.app.common.core.workspace.workspace_dependencies import (
     is_workspace_available,
     is_workspace_owner,
 )
-from studio.app.common.schemas.experiment import (
-    DeleteItem,
-    FetchExptResponse,
-    RenameItem,
-)
+from studio.app.common.schemas.experiment import DeleteItem, RenameItem
 from studio.app.dir_path import DIRPATH
 
 router = APIRouter(prefix="/experiments", tags=["experiments"])
@@ -91,28 +84,6 @@ async def delete_experiment_list(workspace_id: str, deleteItem: DeleteItem):
         return True
     except Exception:
         return False
-
-
-@router.get(
-    "/fetch/{workspace_id}",
-    response_model=FetchExptResponse,
-    dependencies=[Depends(is_workspace_available)],
-)
-async def fetch_last_experiment(workspace_id: str):
-    last_expt_config = ExptUtils.get_last_experiment(workspace_id)
-    if last_expt_config:
-        workflow_config_path = join_filepath(
-            [
-                DIRPATH.OUTPUT_DIR,
-                workspace_id,
-                last_expt_config.unique_id,
-                DIRPATH.WORKFLOW_YML,
-            ]
-        )
-        workflow_config = WorkflowConfigReader.read(workflow_config_path)
-        return FetchExptResponse(**asdict(last_expt_config), **asdict(workflow_config))
-    else:
-        raise HTTPException(status_code=404)
 
 
 @router.get(
